create the galaxy cache if needed
